+++ src/App.jsx
@@ -1,4 +1,4 @@
-import Todo from "./assets/components/Todo";
+import Todo from "./components/Todo";
 
 const App = () => {
   return <Todo />;
